combined interface into class again to stop dependency issue
